% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_sims.R
\name{holdout_sims}
\alias{holdout_sims}
\title{Simulating Imputation to Assess Accuuracy}
\usage{
holdout_sims(dat, network, n.sims, threshold = 0.5, missing_var, holdout.prop)
}
\arguments{
\item{dat}{the dataframe}

\item{network}{object of class bn that specifies the structure of network}

\item{n.sims}{the number of simulations to cycle through}

\item{threshold}{the cuttoff point for "sufficient" probability to impute}

\item{missing_var}{string indicating the variable/feature that you would like to test predictability}

\item{holdout.prop}{proportion of observations to be held out}
}
\value{
a dataframe with 4 slots: holdout, n.sims, thresh_correct, hits_thresh, thresh_correct_prop
}
\description{
Simulating Imputation to Assess Accuuracy
}
\section{Slots}{

\describe{
\item{\code{holdout}}{Number many NAs created)}

\item{\code{n.sims}}{Number of simulations}

\item{\code{thresh_correct}}{Number of predictions that hit the threshold and were correct}

\item{\code{hits_thresh}}{Number of predictions that hit the threshold}

\item{\code{thresh_correct_prop}}{Proportion of predictions that hit the threshold which were correct (thresh_correct/hits_thresh)}
}}

