% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_pred.R
\name{thresh_pred}
\alias{thresh_pred}
\title{Prediction Quality at Different Thresholds}
\usage{
thresh_pred(predict_mat, threshold = seq(0, 1, 0.05), missing_var)
}
\arguments{
\item{predict_mat}{output of m_predict with original dataframe values appended}

\item{threshold}{the vector of cutoff point for "sufficient" probability to impute}

\item{missing_var}{string indicating the variable/feature that you would like to test predictability}
}
\value{
a dataframe with 8 rows and as many columns as number of thresholds

count_match_thresh & prop_match_thresh: observations where prediction and true value match and reaches threshhold

count_nomatch_thresh & prop_nomatch_thresh: observations where prediction and true value do not match and reaches threshhold

count_match_nothresh & prop_match_nothresh: observations where prediction and true value match and fail to reach threshhold

count_nomatch_nothresh & prop_nomatch_nothresh: observations where prediction and true value do not match and fail to reach threshhold
}
\description{
Produces the number and proportion of correctly and incorrectly predicted values based on output of m_predict.
}
\examples{
#m_pred_out is output from an m_predict() command
data(m_pred_out, package = "bnpredict")

#default thresholds
job_thresh <-  thresh_pred(m_pred_out, missing_var="job")
job_thresh

#custom thresholds- 0.5 to 0.9 by 0.05
job_thresh_cut <-  thresh_pred(m_pred_out, missing_var="job", threshold = seq(from = 0.5, to = 0.9, by = 0.05))
job_thresh_cut

#visualize
thresh_viz(job_thresh_cut)

#custom visualization
thresh_viz(job_thresh_cut, colors = c("lightskyblue1", "darkorange1", "darkorange3", "lightskyblue3"),
           main = "My Custom Title", text = T, t.size = 3)
}

