% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_viz.R
\name{thresh_viz}
\alias{thresh_viz}
\title{Visualizing thresh_pred output}
\usage{
thresh_viz(thresh_tab, colors = c("slategray4", "firebrick4", "firebrick1",
  "slategray1"), main = "Imputation Quality at Different Thresholds",
  xlab = "Threshold", ylab = "Prediction Quality Count", pos = "stack",
  text = F, t.size = 5)
}
\arguments{
\item{thresh_tab}{output of thresh_pred()}

\item{colors}{vector of 4 colors for barplot}

\item{main}{title for plot}

\item{xlab}{label of the x axis}

\item{ylab}{label of the y axis}

\item{pos}{type of barplot to produce ("stack", "dodge")}

\item{text}{Logical statement. Include text indicating the count in each segment of the bar}

\item{t.size}{size of the within-plot text (using ggplot sizing)}
}
\value{
a ggplot object that visualizes the number of correctly and incorrectly predicted values
}
\description{
Visualizes the number of correctly and incorrectly predicted values based on output of thresh_pred.
}
\examples{
#m_pred_out is output from an m_predict() command
data(m_pred_out, package = "bnpredict")

#default thresholds
job_thresh <-  thresh_pred(m_pred_out, missing_var="job")
job_thresh

#custom thresholds- 0.5 to 0.9 by 0.05
job_thresh_cut <-  thresh_pred(m_pred_out, missing_var="job", threshold = seq(from = 0.5, to = 0.9, by = 0.05))
job_thresh_cut

#visualize
thresh_viz(job_thresh_cut)

#custom visualization
thresh_viz(job_thresh_cut, colors = c("lightskyblue1", "darkorange1", "darkorange3", "lightskyblue3"),
           main = "My Custom Title", text = T, t.size = 3)
}

