% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_viability.R
\name{value_viability}
\alias{value_viability}
\title{Determining Value Viability}
\usage{
value_viability(dat, network, method = "mle", r, missing_var)
}
\arguments{
\item{dat}{The dataframe}

\item{network}{Object of class bn that specifies the structure of network}

\item{method}{Method of CPT generation for the bn.fit function (i.e. "mle", "bayes")}

\item{r}{Row containing missing data}

\item{missing_var}{String indicating the variable/feature that you would like to test predictability}
}
\value{
a dataframe with 5 slots: observed, observed_prob, match, MPE, MPE_prob
}
\description{
Evaluate the viability of an observed value given the values of all other features for that observation and the the CPT fitted on all other observations in the data.
}
\section{Slots}{

\describe{
\item{\code{observed}}{The observed value of the specified feature}

\item{\code{observed_prob}}{Probability of the observed value of the specified feature}

\item{\code{match}}{Logical value indicating if the value of the MPE matches the observed value}

\item{\code{MPE}}{Most probable value for the specified feature based on the query}

\item{\code{MPE_prob}}{Probability of the MPE}
}}
\examples{
data(dat, package = "bnpredict")
net <- bnlearn::model2network("[compensation|age:full_time][age][job|compensation:full_time]
                      [full_time][cluster|job:compensation:age:full_time]")

#evaluating viability of compensation in the first row of dat
value_viability(dat= dat, r= 15, network = net,  missing_var = "compensation")

#using plyr to evaluate viability of first 200 rows of dat for compensation and job title
comp_out <- plyr::adply(dat[1:200,], 1, function(x) value_viability(dat = dat, network = net,
                                                              r= max(as.numeric(rownames(x))),
                                                              missing_var = "compensation"), .progress = "text")

job_out <- plyr::adply(dat[1:200,], 1, function(x) value_viability(dat = dat, network = net,
                                                             r= max(as.numeric(rownames(x))),
                                                             missing_var = "job"), .progress = "text")
}

